refactor: move file to dedicated pkce namespace
